parser: fix keyword parsing
It fundamentally needs to use `trying' because when we parse an
identifier, such as 'index' we try to parse a keyword first to make
sure that the user isn't trying to use a keyword as an identifier.
`in` is a keyword, so when parsing 'index' we would _succeed_ in
parsing 'in' and then fail because a keyword can't be followed by any
alphanumeric characters. The parser was choking on this exact example.