closure-convert tm declarations; eta-expand constructors
Closure conversion has been extended to process programs (i.e. lists of
declarations)

Type declarations are for the most part skipped. We just keep track of
the arities of constructors.
We unconditionally convert the body of every term definition, and then
see whether it's a degenerate MkClo (with an empty environment
renaming). We simply change this into a call to the synthetic function
`f` bundled with the MkClo and take note that this term is a function
with a statically-known arity.

The reason we keep track of the constructor arities is so that we can
eta-expand them during closure conversion. When we see a constructor
application `c tS` we might need to extend `tS` with variables and wrap
the resulting application in a corresponding number of `fun ! -> ...`
abstractions. Such functions are then closure-converted.

This is nice because in the bytecode compilation later we will be able
to assume that constructors are saturated. I wonder if it's worth doing
the same for calls to known functions (via Ref)? It would probably
simplify the bytecode compilation because I wouldn't need to figure out
whether calls are saturated later.