add eta-expansion of known function calls and prims
I realized that we could eta expand a lot more than just constructors.
This way, any call to a known function of a known arity (including
primitives) is eta-expanded as well. This will make it easier in the
compiler because it can assume that any App to a Ref, Const, or Prim has
a spine _at least_ as long as the arity of that head.
That "at least" is very important: it is possible that the spine be
_longer_ than the known arity, for instance due to a polymorphic call
instantiating a return type to a function type, or because a Ref ends up
referring to a closure.

I'm not sure in these closure cases whether to statically consider the
closure to be a function (of a >0 arity) or whether to consider it to
have a zero arity.