understand what arity 0 finally means
When a declaration has arity 0, that means that we need to actually
evaluate the RHS as part of running the program. When arity n > 0, we
instead have a genuine function body (perhaps a closure body, or a pure
function).

Compilation will then process all the declarations and collect the code
for each arity 0 decl into the main program text. Arity n>0 declarations
will form functions and do not end up in the main program text.