diagnose eta-expansion bug; refactor
Remove `compilerCommon.ml` and rename it to `programInfo.ml`.
Add a function to that module to collect program info from an internal
syntax program signature, and some helpers for looking up refs and ctor
info. Such functions were present in both close.ml and compile.ml so now
both of these modules are slightly simplified since they don't need to
know as much about the internals of the ProgramInfo struct.